+++ modules/eslint-plugin/spec/rules/store/prefer-one-generic-in-create-for-feature-selector.spec.ts
@@ -1,4 +1,8 @@
-import type { ESLintUtils, TSESLint } from '@typescript-eslint/utils';
+import type { ESLintUtils } from '@typescript-eslint/utils';
+import type {
+  InvalidTestCase,
+  ValidTestCase,
+} from '@typescript-eslint/rule-tester';
 import * as path from 'path';
 import rule, {
   preferOneGenericInCreateForFeatureSelector,
@@ -8,15 +12,14 @@ import { ruleTester, fromFixture } from '../../utils';
 
 type MessageIds = ESLintUtils.InferMessageIdsTypeFromRule<typeof rule>;
 type Options = ESLintUtils.InferOptionsTypeFromRule<typeof rule>;
-type RunTests = TSESLint.RunTests<MessageIds, Options>;
 
-const valid: () => RunTests['valid'] = () => [
+const valid: () => (string | ValidTestCase<Options>)[] = () => [
   `const createFeatureSelector = test('feature-state')`,
   `const featureOk = createFeatureSelector('feature-state')`,
   `const featureOk1 = createFeatureSelector<FeatureState>('feature-state')`,
 ];
 
-const invalid: () => RunTests['invalid'] = () => [
+const invalid: () => InvalidTestCase<MessageIds, Options>[] = () => [
   fromFixture(
     `
 const featureNotOk = createFeatureSelector<GlobalState, FeatureState>('feature-state')