+++ modules/eslint-plugin/spec/rules/store/select-style.spec.ts
@@ -1,13 +1,16 @@
-import type { ESLintUtils, TSESLint } from '@typescript-eslint/utils';
+import type { ESLintUtils } from '@typescript-eslint/utils';
+import type {
+  InvalidTestCase,
+  ValidTestCase,
+} from '@typescript-eslint/rule-tester';
 import * as path from 'path';
 import rule, { SelectStyle } from '../../../src/rules/store/select-style';
 import { ruleTester, fromFixture } from '../../utils';
 
 type MessageIds = ESLintUtils.InferMessageIdsTypeFromRule<typeof rule>;
 type Options = readonly ESLintUtils.InferOptionsTypeFromRule<typeof rule>[0][];
-type RunTests = TSESLint.RunTests<MessageIds, Options>;
 
-const validConstructor: () => RunTests['valid'] = () => [
+const validConstructor: () => (string | ValidTestCase<Options>)[] = () => [
   `
 import { Store } from '@ngrx/store'
 
@@ -90,7 +93,7 @@ class Ok9 {
   },
 ];
 
-const validInject: () => RunTests['valid'] = () => [
+const validInject: () => (string | ValidTestCase<Options>)[] = () => [
   `
 import { Store } from '@ngrx/store'
 import { inject } from '@angular/core'
@@ -162,7 +165,7 @@ class Ok17 {
   },
 ];
 
-const invalidConstructor: () => RunTests['invalid'] = () => [
+const invalidConstructor: () => InvalidTestCase<MessageIds, Options>[] = () => [
   fromFixture(
     `
 import { select, Store } from '@ngrx/store'
@@ -333,7 +336,7 @@ class NotOk5 {
   ),
 ];
 
-const invalidInject: () => RunTests['invalid'] = () => [
+const invalidInject: () => InvalidTestCase<MessageIds, Options>[] = () => [
   fromFixture(
     `
 import { select, Store } from '@ngrx/store'