+++ projects/example-app/src/app/books/components/book-preview-list.component.ts
@@ -16,6 +16,7 @@ import { Book } from '@example-app/books/models';
       }
     `,
   ],
+  standalone: false,
 })
 export class BookPreviewListComponent {
   @Input() books!: Book[];