+++ projects/example-app/src/app/books/components/book-preview.component.ts
@@ -76,6 +76,7 @@ import { Book } from '@example-app/books/models';
       }
     `,
   ],
+  standalone: false,
 })
 export class BookPreviewComponent {
   @Input() book!: Book;