+++ projects/example-app/src/app/core/components/layout.component.ts
@@ -14,5 +14,6 @@ import { Component } from '@angular/core';
       }
     `,
   ],
+  standalone: false,
 })
 export class LayoutComponent {}